requirements.txt
 
pypdf2-1.27.6

pdfx-1.4.1